[Chore] Renamed interfaces file to interfaces.ts
