+++ src/classLibrary/interfaces/interfaces.ts
